requirements.txt
django==1.7.4
openpyxl
python-dateutil
django-report-utils>=0.2.9
djangorestframework==3.0.4
psycopg2
django-extensions
Werkzeug
celery[redis]
